subscription implementation

+++ code/PersistenceTier/src/main/java/via/sep3/persistencetier/database/reservation/ReservationRepository.java
@@ -8,7 +8,7 @@ import via.sep3.persistencetier.database.seller.Seller;
 
 import java.util.stream.Stream;
 
-public interface ReservationRepository extends JpaRepository<Reservation, Long> {
+public interface ReservationRepository extends JpaRepository<Reservation, Integer> {
